+++ app/models/research_session/steps.rb
@@ -12,7 +12,7 @@ class ResearchSession
       methodologies: [:other_methodology, methodologies: []],
       recording:     [:other_recording_method, recording_methods: []],
       data:          [:shared_with, :shared_duration, :shared_use],
-      time_equipment: [:start_datetime, :duration, :participant_equipment],
+      time_equipment: [:start_datetime, :duration, :location, :participant_equipment],
       expenses:      [:travel_expenses_limit, :food_expenses_limit, :other_expenses_limit,
                       :receipts_required, :food_provided],
       incentive:     [:incentive, :payment_type, :incentive_value]