+++ db/migrate/20171017154810_add_location_to_time_equipment.rb
@@ -1,0 +1,5 @@
+class AddLocationToTimeEquipment < ActiveRecord::Migration[5.1]
+  def change
+    add_column :research_sessions, :location, :string
+  end
+end