+++ features/step_definitions/question_steps.rb
@@ -71,6 +71,8 @@ When(/^I provide full session details for every step$/) do
   end
   @session_duration = '5 minutes'
   fill_in 'How long will the session be? (optional)', with: @session_duration
+  @session_location = 'Rockford House, Leeds'
+  fill_in 'Where will the session be? (optional)', with: @session_location
   @what_to_bring = 'Nothing'
   fill_in 'What do participants need to bring? (optional)', with: @what_to_bring
   Percy::Capybara.snapshot(page, name: :time_equipment)